LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain-expected.txt


PASS sanity checks 
PASS current document 
FAIL new document assert_equals: new document has empty domain expected "" but got "localhost"
